+++ src/index.ts
@@ -1,23 +1,7 @@
-import { Context, Middleware } from '@curveball/core';
-import highlight from 'highlight.js';
-
-/**
- * Options that may be passed to the middleware.
- */
-type Options = {
-  title?: string
-  stylesheets?: string[]
-};
-
-/**
- * Options after clean-up.
- *
- * Basically the same but nothing is optional.
- */
-type SureOptions = {
-  title: string,
-  stylesheets: string[]
-};
+import { Middleware } from '@curveball/core';
+import generateHtmlIndex from './html-index';
+import serveAsset from './serve-asset';
+import { NavigationLinkMap, Options, SureOptions } from './types';
 
 const parsedContentTypes = [
   'application/json',
@@ -25,22 +9,56 @@ const parsedContentTypes = [
   'application/problem+json',
 ];
 
+/*
+ * Wanted links support
+ *
+ * source:
+ * https://www.iana.org/assignments/link-relations/link-relations.xhtml
+ *
+ * - about
+ * - alternate
+ * - collection
+ * - help
+ * - item
+ * - stylesheet
+ * - up
+ * - via
+ *
+ * source:
+ * http://microformats.org/wiki/existing-rel-values
+ *
+ * - code-repository
+ * - icon
+ */
+
+const defaultNavigationLinks: NavigationLinkMap = {
+  'author': true,
+  'create-form': true,
+  'edit': true,
+  'edit-form' : {
+    // alias this
+    icon: 'icons/edit.svg',
+  },
+  'home': true,
+  'next': true,
+  'previous': true,
+  'search': true,
+};
 
 export default function browser(options?: Options): Middleware {
 
-  if (typeof options === 'undefined') {
-    options = {};
-  }
-
-  if (!options.title) {
-    options.title = 'HAL Browser';
-  }
-
-  if (!options.stylesheets) {
-    options.stylesheets = [];
-  }
+  const newOptions = normalizeOptions(options);
 
   return async (ctx, next) => {
+
+    if (newOptions.serveAssets && ctx.request.path.startsWith('/_hal-browser/assets/')) {
+      return serveAsset(ctx);
+    }
+
+    // Don't do anything if the raw format was requested
+    if ('_browser-raw' in ctx.request.query) {
+      return next();
+    }
 
     // Check to see if the client even wants html.
     if (!ctx.request.accepts('text/html')) {
@@ -61,108 +79,69 @@ export default function browser(options?: Options): Middleware {
     // This is useful if the client submitted a lower q= score for text/html
     if (ctx.request.accepts('text/html', ...parsedContentTypes) === 'text/html') {
       ctx.response.headers.set('Content-Type', 'text/html');
-      generateHtmlIndex(ctx, ctx.response.body, <SureOptions> options);
+      generateHtmlIndex(ctx, ctx.response.body, newOptions);
     }
 
   };
 
 }
 
-function generateHtmlIndex(ctx: Context, body: any, options: SureOptions) {
 
-  const jsonBody = syntaxHighlightJson(body);
-  const links = generateLinks(body);
+/**
+ * This function does a whole bunch of cleanup of the options object, so
+ * everything else can do less work.
+ *
+ * This makes the rest of the source simpler, and also saves time because it
+ * only happens once.
+ */
+function normalizeOptions(options: Options): SureOptions {
 
-  const stylesheets = options.stylesheets.map(ss => {
-    return `    <link rel="stylesheet" href="${h(ss)}" type="text/css" />\n`;
-  }).join('');
-
-  ctx.response.body = `
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>${h(options.title)}</title>
-${stylesheets}
-  </head>
-  <body>
-    <header>
-      <h1>${h(options.title)}</h1>
-    </header>
-
-    <main>
-      ${links}
-
-      <h2>Body</h2>
-      <code class="hljs"><pre>${jsonBody}</pre></code>
-
-    </main>
-
-  </body>
-</html>
-`;
-
-}
-
-function h(input: string): string {
-
-  const map: { [s: string]: string } = {
-    '&' : '&amp;',
-    '<' : '&lt;',
-    '>' : '&gt;',
-    '"' : '&quot'
-  };
-
-  return input.replace(/&<>"/g, s => map[s]);
-
-}
-
-function syntaxHighlightJson(body: any): string {
-
-  return highlight.highlight('json', JSON.stringify(body, undefined, '  ')).value;
-
-}
-
-function generateLinks(body: any): string {
-
-  if (!body._links) { return ''; }
-
-  let linkHtml = '';
-
-  for (const rel of Object.keys(body._links)) {
-
-    const links =
-      Array.isArray(body._links[rel]) ?
-      body._links[rel] :
-      [body._links[rel]];
-
-    const linkCount = links.length;
-    let first = true;
-
-    for (const link of links) {
-
-      linkHtml += '<tr>';
-      if (first) {
-        linkHtml += `<td rowspan="${linkCount}">${h(rel)}</td>`;
-        first = false;
-      }
-      linkHtml += `<td><a href="${h(link.href)}">${h(link.href)}</a></td>`;
-      linkHtml += '<td>' + (link.title ? h(link.title) : '') + '</td>';
-      linkHtml += '</tr>\n';
-
-    }
-
-
+  if (typeof options === 'undefined') {
+    options = {};
   }
 
-  return `
-    <h2>Links</h2>
-    <table>
-      <tr>
-        <th>Relationship</th><th>Url</th><th>Title</th>
-      </tr>
-      ${linkHtml}
-    </table>
-  `;
+  const defaults = {
+    title: 'HAL Browser',
+    stylesheets: [
+      'css/main.css',
+      'css/solarized-dark.css',
+    ],
+    defaultLinks: [
+      {
+        href: '/',
+        rel: 'home'
+      }
+    ],
+    hiddenRels: [
+      'self',
+      'curies',
+    ],
+    assetBaseUrl: '/_hal-browser/assets/',
+    serveAssets: true,
+  };
+
+  const tmpNavLinks = Object.assign(
+    defaultNavigationLinks,
+    options.navigationLinks === undefined ? {} : options.navigationLinks
+  );
+
+  options.navigationLinks = {};
+
+  for (const navLinkRel of Object.keys(tmpNavLinks)) {
+
+    const navLink = tmpNavLinks[navLinkRel];
+    if (navLink === null) {
+      continue;
+    }
+    if (navLink === true) {
+      options.navigationLinks[navLinkRel] = {
+        defaultTitle: navLinkRel,
+      };
+    } else {
+      options.navigationLinks[navLinkRel] = navLink;
+    }
+  }
+
+  return <SureOptions> Object.assign(defaults, options);
 
 }
-