+++ changelog.md
@@ -1,8 +1,12 @@
-0.1.0 (????-??-??)
+0.2.0 (2018-07-07)
 ==================
 
-* Some colorscheme refresh. Still not great. #notadesigner
-
+* Automatically generate navigation icons.
+* Hide uninteresting links.
+* Automatically detect title of resource.
+* Colorscheme refresh. Still not great. #notadesigner
+* Lots of tiny tweaks and bugfixes.
+* Serve `.svg` and `.css` files out of the box.
 
 0.0.2 (2018-07-02)
 =================